week/week2/day4/exercises_xp/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
psycopg2-binary==2.9.10
    # via exercises-xp (pyproject.toml)
python-dotenv==1.1.1
    # via exercises-xp (pyproject.toml)
